esp-zigbee-sdk: Provide zboss APIs to fix github issues(70cea8e)

See merge request espressif/esp-zigbee-sdk!36

+++ components/esp-zigbee-lib/include/zcl/esp_zigbee_zcl_command.h
@@ -66,8 +66,8 @@ typedef struct esp_zb_zcl_config_report_cmd_s {
     uint16_t clusterID;                             /*!< Cluster ID to report */
     uint16_t attributeID;                           /*!< Attribute ID to report */
     uint8_t  attrType;                              /*!< Attribute type to report refer to zb_zcl_common.h zcl_attr_type */
-    uint8_t  min_interval;                          /*!< Minimum reporting interval */
-    uint8_t  max_interval;                          /*!< Maximum reporting interval */
+    uint16_t  min_interval;                         /*!< Minimum reporting interval */
+    uint16_t  max_interval;                         /*!< Maximum reporting interval */
     uint16_t reportable_change;                     /*!< Minimum change to attribute will result in report */
 } esp_zb_zcl_config_report_cmd_t;
